+++ ichnaea/country.py
@@ -11,9 +11,10 @@ from country_bounding_boxes import (
     countries,
     country_subunits_by_iso_code,
 )
-
+import mobile_codes
 from six import string_types
 
+from ichnaea.constants import ALL_VALID_COUNTRIES
 from ichnaea import geocalc
 
 _bbox_cache = []
@@ -28,6 +29,8 @@ def _fill_bbox_cache():
         iso3 = _best_guess_iso_3(country)
         if not (iso2 and iso3):
             continue
+        if iso2 not in ALL_VALID_COUNTRIES:
+            continue
         (lon1, lat1, lon2, lat2) = country.bbox
         radius = geocalc.distance(lat1, lon1, lat2, lon2) / 2.0
         cached.append(Subunit(bbox=country.bbox,
@@ -36,6 +39,18 @@ def _fill_bbox_cache():
                               radius=radius))
     # sort by largest radius first
     return list(sorted(cached, key=attrgetter('radius'), reverse=True))
+
+
+def countries_for_mcc(mcc):
+    """
+    Return a list of :class:`mobile_codes.Country` instances matching
+    the passed in mobile country code.
+
+    The return list is filtered by the set of recognized ISO 3166 alpha2
+    codes present in the GENC dataset.
+    """
+    countries = mobile_codes.mcc(str(mcc))
+    return [c for c in countries if c.alpha2 in ALL_VALID_COUNTRIES]
 
 
 def country_for_location(lat, lon):
@@ -56,12 +71,14 @@ def country_for_location(lat, lon):
     return None
 
 
-def country_matches_location(lat, lon, country_code, margin=0.0):
+def country_matches_location(lat, lon, code, margin=0.0):
     """
     Return whether or not a given (lat, lon) pair is inside one of the
     country subunits associated with a given alpha2 country code.
     """
-    for country in country_subunits_by_iso_code(country_code):
+    if code not in ALL_VALID_COUNTRIES:
+        return False
+    for country in country_subunits_by_iso_code(code):
         (lon1, lat1, lon2, lat2) = country.bbox
         if lon1 - margin <= lon and lon <= lon2 + margin and \
            lat1 - margin <= lat and lat <= lat2 + margin: